renaming repo errors
