+++ src/infrastructure/repositories/redis-short-url.repository.unit.test.ts
@@ -1,7 +1,7 @@
 import { Redis } from 'ioredis'
 
 import { ShortUrl } from '@domain/entities/short-url.js'
-import { CodeExistsError } from '@infrastructure/errors/code-exists.error.js'
+import { CodeExistsError } from '@infrastructure/errors/repository.error.js'
 import { RedisShortUrlRepository } from '@infrastructure/repositories/redis-short-url.repository.js'
 
 describe('RedisShortUrlRepository', () => {
